tests/wpt/tests/css/css-fonts/parsing/font-variant-position-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: getComputedStyle().fontVariantPosition</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-variant-position-prop">
<meta name="assert" content="font-variant-position computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('font-variant-position', 'normal');
test_computed_value('font-variant-position', 'sub');
test_computed_value('font-variant-position', 'super');
</script>
</body>
</html>
